Fixed vertical lines in display.
Because I was multiplying the infinity value in tDelta by zero (and infinity * 0 = NaN), I was getting some number += NaN = NaN. Then in the next mask check, the infinity axis evaluates to false and the ray never moves in that direction. The fix is eithen not to multiply infinity by zero, or to treat NaN as infinity. The former was easier to implement.